Feat/ios 5 (#5)
* exercise controller

update

* cleanup

+++ code/backend/src/application/qb8s.net.OptiFit.CQRS/Commands/Exercise/DeleteExerciseCommand.cs
@@ -6,21 +6,26 @@ using qb8s.net.OptiFit.Persistence.Context;
 
 namespace qb8s.net.OptiFit.CQRS.Commands.Exercise;
 
-public record DeleteExerciseCommand(Guid UserId, Guid Id) : IRequest;
+public record DeleteExerciseCommand(Guid UserId, Guid Id) : IRequest<Guid>;
 
 public class DeleteExerciseCommandHandler(
     ILogger<DeleteExerciseCommandHandler> logger,
     IMapper mapper,
-    ApplicationDbContext dbContext) : IRequestHandler<DeleteExerciseCommand>
+    ApplicationDbContext dbContext) : IRequestHandler<DeleteExerciseCommand, Guid>
 {
-    public async Task Handle(DeleteExerciseCommand request, CancellationToken cancellationToken)
+    public async Task<Guid> Handle(DeleteExerciseCommand request, CancellationToken cancellationToken)
     {
-        //check for user permissions / (new Guid("275cfdca-c686-4ea1-80b1-f2425b1602c5")
+        var user = await dbContext.Users.Include(user => user.UserRole)
+            .FirstOrDefaultAsync(user => user.Id == request.UserId, cancellationToken);
+        if (user == null) throw new Exception("User not found");
+        if (!user.UserRole.Name.Contains("Admin"))
+            throw new Exception("User does not have permission to delete exercise");
         logger.LogInformation("Delete Exercise Command : {@Id}", request.Id);
         var entity = await dbContext.Exercises.FirstOrDefaultAsync(e => e.Id == request.Id, cancellationToken);
         if (entity == null) throw new Exception("Exercise not found");
 
         dbContext.Exercises.Remove(entity);
         await dbContext.SaveChangesAsync(cancellationToken);
+        return entity.Id;
     }
 }
